SysUI: Use VectorDrawable's for rendering battery
The current implementation of BatteryMeterView uses arrays of points
that are drawn programmatically.  This limits the ability to overlay
or theme the battery.  By utilizing VectorDrawables and
AnimatedVectorDrawables we can achieve the same look and feel of
stock while allowing for more unique battery meters.

Change-Id: I16a8a949a92827a9e8c1b528981031b6830e97cd

Squashed:

Fix drawing of VectorDrawables that are rotated by 90 or 270 degrees.
The SCALE_X / SCALE_Y values of rotated matrices are set to the cosine
of the rotation angle, which is 0 for drawables that are rotated by 90
or 270 degrees. Map the matrix onto the bounds rectangle and use the
resulting size to determine the scaled size instead.

Change-Id: I6cc0f0b5569d43434c829cb5f6e5f848cad995ec

Add StopMotionVectorDrawable
The StopMotionVectorDrawable provides control over the position
within an animation set of the warpped AnimatedVectorDrawable.

- Moved StopMotionVectorDrawable to graphics/java/android/graphics/drawable
as that would be where Google would put it if they decided to do something
like this.

Thanks @elginsk8r for the suggestion

Change-Id: I40c93074cfec0605faffeb29a0e11f2c89db58e0

SystemUI: Set circle battery path as untranslatable
Change-Id: I93870985fc3953fc49aeaf9f169553685d7a202a

SysUI: Adjust battery dimensions for better look
Make some minor tweaks to the new vector based batter meter to get
it even closer to stock look.

Change-Id: I80fcdefb2578352bef5ca35ad902556b204de322

SysUI: Use linear interpolator for battery level
By default, ObjectAnimator uses the AccelerateDecelerate interpolator
which causes the battery meter to show the incorrect level.

Explicitly setting the interpolator to linear fixes this issue.

Change-Id: I06e2c743aaa6e0560f0fe8fe27377427becb11c7
TICKET: NIGHTLIES-2985

SysUI: Allow styling of battery text/charge blend mode
Not all battery meter modes are created equally and sometimes you
need to draw the text or charge indicator using a different blend
mode than XOR.  This patch adds a new attribute to the battery
meter style which allows the blend mode to be defined per battery
style.

Change-Id: I3d2eb38828abc7e5d93337fa815a414675df43a1
TICKET: NIGHTLIES-2981

SysUI: Fix coloring of frame and bolt
This patch should address issues with the frame and bolt being
incorrectly colored in dark/light scenarios.  We also keep track
of the darkness intensity level so that it can be set when changing
meter modes.

Change-Id: Iaec8b88d39bbb7f6dbcbbf078c91d20b54ab29c8